requirements.txt
# Framework web
fastapi
uvicorn[standard]
python-multipart
pydantic-settings

# OCR y procesamiento de imágenes
pytesseract
easyocr
paddleocr  # Nuevo OCR principal
ocrmypdf
opencv-python
pillow
pdf2image
pymupdf  # (instala fitz)

# Donut OCR (opcional, para casos especiales)
transformers  # requerido por Donut
sentencepiece  # requerido por Donut
donut-python  # modelo OCR-free

# Modelos LLM y NLP
# llama-cpp-python  # (lo dejaste comentado)
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0

# Utilidades generales
numpy<2
pandas
requests
python-dotenv
scipy

# Testing
pytest
httpx
